refactor: split stats persistence helpers
